fix return type of Nouns token interface

+++ packages/nouns-contracts/contracts/governance/NounsDAOInterfaces.sol
@@ -436,5 +436,5 @@ interface INounsDAOExecutor {
 interface NounsTokenLike {
     function getPriorVotes(address account, uint256 blockNumber) external view returns (uint96);
 
-    function totalSupply() external view returns (uint96);
+    function totalSupply() external view returns (uint256);
 }
